#16: Remove unnecessary setters in ApplicationLog classes; Move HexUtils methods to proper classes.
